Updated the msvc compiler script
I have updated the msvc compiler script and removed the use of msbuild
and the conversion of the solution files.

All of the extension modules and DLLs are not built internally to the
setup script using distutils. This makes a centralized location for
maintaining the build code for the extensions and DLLs.

The script for the c compiler now supports a wider range of msvc versions
from 9.0 all the way to the most recent) supports all Windows SDK's as
well.